+++ packages/mantra-dex-std/src/pool_manager.rs
@@ -115,6 +115,16 @@ pub struct PoolStatus {
     pub withdrawals_enabled: bool,
 }
 
+impl Default for PoolStatus {
+    fn default() -> Self {
+        PoolStatus {
+            swaps_enabled: true,
+            deposits_enabled: true,
+            withdrawals_enabled: true,
+        }
+    }
+}
+
 /// The contract configuration.
 #[cw_serde]
 pub struct Config {
@@ -124,8 +134,6 @@ pub struct Config {
     pub farm_manager_addr: Addr,
     /// How much it costs to create a pool. It helps prevent spamming of new pools.
     pub pool_creation_fee: Coin,
-    //  Whether or not swaps, deposits, and withdrawals are enabled
-    pub feature_toggle: FeatureToggle,
 }
 
 #[cw_serde]
@@ -221,8 +229,8 @@ pub enum ExecuteMsg {
         farm_manager_addr: Option<String>,
         /// The new fee that must be paid when a pool is created.
         pool_creation_fee: Option<Coin>,
-        /// The new feature toggles of the contract, allowing fine-tuned
-        /// control over which operations are allowed.
+        /// Toggles features for a given pool, allowing fine-tuned
+        /// control over which operations are allowed, i.e. swap, deposits, withdrawals
         feature_toggle: Option<FeatureToggle>,
     },
 }
@@ -364,12 +372,14 @@ pub struct ReverseSimulationResponse {
 /// Pool feature toggle, can control whether swaps, deposits, and withdrawals are enabled.
 #[cw_serde]
 pub struct FeatureToggle {
+    /// The identifier of the pool to toggle the status of.
+    pub pool_identifier: String,
     /// Whether or not swaps are enabled
-    pub withdrawals_enabled: bool,
+    pub withdrawals_enabled: Option<bool>,
     /// Whether or not deposits are enabled
-    pub deposits_enabled: bool,
+    pub deposits_enabled: Option<bool>,
     /// Whether or not swaps are enabled
-    pub swaps_enabled: bool,
+    pub swaps_enabled: Option<bool>,
 }
 
 /// The response for the `SimulateSwapOperations` query.